requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
google-generativeai==0.3.2
pydantic[dotenv]==2.5.2
